Add initial implementation
This is by no means perfect and I'm particularly worried about how
timezones are handles considering toggl exports appear to use local
time; I can't tell if it's the timezone of the entry or the export
because I haven't been traveling. But my historical data, mostly from
a previous job, is not that important to me and I think it is time to
move on.